requirements.txt
python-dotenv==1.0.0
pyTelegramBotAPI==4.14.0
instagrapi==2.1.2
PyYAML==6.0.1
pydantic==2.7.1
ipython==8.17.2
Pillow==8.1.1